If the source file doesn't exist, don't copy it.

+++ pisync/pisync-receiver.sh
@@ -4,4 +4,7 @@
 DEST_PATH="/etc/pihole/hosts/custom.list"
 SOURCE_PATH="/opt/pisync/custom.list"
 
-cp -f $SOURCE_PATH $DEST_PATH
+
+if [ -f $SOURCE_PATH ]; then
+    cp -f $SOURCE_PATH $DEST_PATH
+fi
